Verify that callback matches permitted prefix
OAuth callback URLs are only allowed to specify a custom callback
if they are prefixed with the registered callback, which in our
case is always the /oauth_callback one that we provide. This adds
a check for this when a custom callback URL is being supplied in
the query string, and if there's no match it defaults to what is
set in the OAuth consumer.

https://phabricator.wikimedia.org/T230085

Bug: T230085