Change Intuition::msg() to check all domains for messages
The bundle comes with some i18n messages, and they've not been usable
via Twig because we've been setting the Intuition domain to only
one value and the bundle's messages are under the 'toolforge' domain.

This change is to loop through all registered domains (i.e. two of
them, in the usual case), and return the first matched message.

Bug: T273840