SshCommand: Only force username when requested
This removes the use of the result of `whoami` in determining the SSH
user to run the command as. This means a provided username will be
respected, follow by a config-defined `User`, in that order.

This fixes SshCommand ignoring the config-defined `User` and prevents
the use of `whoami` on Windows, which would return a `machine\user`
format instead of just `user` (the default when no explicit username
is provided when running the built-in OpenSSH client).