Add support to connect to a Trove database
Several clients (or would-be clients) connect to a WMCS Trove database,
as opposed to using tools-db. This is customary for any VPS tool.

Theoretically we could just allow the --toolsdb to take a hostname, and
use that and the TOOLSDB_ env variables to accomplish the same thing,
but tools could need to connect to both.